Fix ncurses path mismatch introduced in a9630e3
Currently, macOS comes with the ncurses library installed at
/usr/lib/libncurses.dylib; however with commit a9630e3 in late 2015,
gptfdisk has added a line to Makefile.mac that references
`/opt/local/lib/libncurses.a`. For most OS X/macOS users, this breaks
the build because the file does not ship with macOS.

For more info on the issue, see the relevant discussion on the
Homebrew project. [1]

The commit at hand changes that reference to
/usr/lib/libncurses.dylib, which has shipped with OS X/macOS for many
years, and is guaranteed to exist.

[1]: https://github.com/Homebrew/legacy-homebrew/pull/46398#issuecomment–160473325